Refacator: 공유스페이스 dto 수정
