+++ src/main/java/com/fastcampus/jober/domain/spacewall/repository/SpaceWallRepository.java
@@ -4,6 +4,10 @@ import com.fastcampus.jober.domain.spacewall.domain.SpaceWall;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface SpaceWallRepository extends JpaRepository<SpaceWall, Long> {
+
+    Optional<SpaceWall> findByIdAndDeletedAtIsNull(Long id);
 }
